A bunch of acceptance tests

+++ app/controllers/answers_controller.rb
@@ -1,15 +1,18 @@
 class AnswersController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   expose :question, id: -> { params[:question_id] }
   expose :answer
   expose :answers, -> { Answer.all }
 
   def create
     @answer = question.answers.new(answer_params)
+    @answer.user = current_user
 
     if @answer.save
-      redirect_to @answer
+      redirect_to @answer.question, notice: 'Your answer has been successfully created.'
     else
-      render :new
+      render 'questions/show'
     end
   end
 
@@ -22,13 +25,22 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    answer.destroy
-    redirect_to question_answers_path(question)
+    if answer.user == current_user
+      answer.destroy
+      flash[:notice] = 'Your answer has been successfully removed'
+    else
+      flash[:alert] = 'You are not permitted to delete others\' answers'
+    end
+    redirect_to question_path(answer.question)
   end
 
   private
 
   def answer_params
     params.require(:answer).permit(:body)
+  end
+
+  def rescue_from_record_not_found(error)
+    render plain: error.message
   end
 end
